[#51] 코디 수동 추가 api 적용
#51

+++ Codive/Features/LookBook/Presentation/ViewModel/AddBeforeCodiViewModel.swift
@@ -21,14 +21,14 @@ final class AddBeforeCodiViewModel: ObservableObject {
     
     private let navigationRouter: NavigationRouter
     private let beforeCodiUseCase: BeforeCodiUseCase
-    private let coordinateId: Int
+    private let coordinateId: Int64
     
     // MARK: - Initializer
     
     init(
         navigationRouter: NavigationRouter,
         beforeCodiUseCase: BeforeCodiUseCase,
-        coordinateId: Int
+        coordinateId: Int64
     ) {
         self.navigationRouter = navigationRouter
         self.beforeCodiUseCase = beforeCodiUseCase
